requirements.txt
pandas
Faker
psycopg2-binary
pytest